requirements.txt
opencv-python
pillow
pynput
pyperclip
requests
numpy
python-dotenv